Flood fill based tool

+++ contour-patcher/utilities/image_utils.py
@@ -23,9 +23,9 @@ def crop(img: Mat, x0, y0, x1, y1) -> Mat:
     return img[y0:y1, x0:x1]
 
 
-def apply_zoom(x, y, zoom_level) -> tuple[float, float]:
-    return x * zoom_level, y * zoom_level
+def apply_zoom(x, y, zoom_level) -> tuple[int, int]:
+    return int(x * zoom_level), int(y * zoom_level)
 
 
-def unapply_zoom(x, y, zoom_level) -> tuple[float, float]:
-    return x / zoom_level, y / zoom_level
+def unapply_zoom(x, y, zoom_level) -> tuple[int, int]:
+    return int(x / zoom_level), int(y / zoom_level)
